// minera.frontend/projects/app-admin/src/app/main/models/sku-model.ts
import { AgrupamentoModel } from './agrupamento-model';

export class SkusModel {
    codigoAgrupamento: number;
    nome: string;
    letra: any;
    departamentos: any;
    categorias: any;
    subCategorias: any;
    segmentos: any;
    associacaoNaBusca: boolean;
    filtroAtributoCadastrado: any[];
}